fix bad import + cyclic dependencies
