+++ src/components/TopBar/TopBar.test.tsx
@@ -12,11 +12,12 @@ import { IntlProvider } from 'react-intl';
 import { red } from '@mui/material/colors';
 import { createTheme, ThemeProvider } from '@mui/material';
 import { afterEach, beforeEach, expect, it } from '@jest/globals';
-import TopBar, { LANG_ENGLISH } from './TopBar';
+import TopBar from './TopBar';
 import top_bar_en from '../translations/top-bar-en';
 import { AppMetadataCommon } from '../../services';
 
 import PowsyblLogo from '../images/powsybl_logo.svg?react';
+import { LANG_ENGLISH } from '../../utils/language';
 
 let container: Element;
