+++ src/components/filter/criteria-based/criteria-based-filter-edition-dialog.tsx
@@ -28,7 +28,8 @@ import { FilterType } from '../constants/filter-constants';
 import FetchStatus from '../../../utils/FetchStatus';
 import { exploreSvc } from '../../../services/instances';
 import FilterForm from '../filter-form';
-import { GsLangUser } from '../../TopBar/TopBar';
+
+import { GsLangUser } from '../../../utils/language';
 
 export type SelectionCopy = {
     sourceItemUuid: UUID | null;
