+++ src/utils/EquipmentType.ts
@@ -6,7 +6,8 @@
  */
 
 import { Theme } from '@mui/material';
-import { LIGHT_THEME } from '../components/TopBar/TopBar';
+
+import { LIGHT_THEME } from './theme';
 
 export const TYPE_TAG_MAX_SIZE = '90px';
 export const VL_TAG_MAX_SIZE = '100px';
