+++ src/utils/language.ts
@@ -5,15 +5,14 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import {
-    GsLang,
-    GsLangUser,
-    LANG_ENGLISH,
-    LANG_FRENCH,
-    LANG_SYSTEM,
-} from '../components/TopBar/TopBar';
+export const LANG_SYSTEM = 'sys';
+export const LANG_ENGLISH = 'en';
+export const LANG_FRENCH = 'fr';
 
 const supportedLanguages = [LANG_FRENCH, LANG_ENGLISH];
+
+export type GsLangUser = typeof LANG_ENGLISH | typeof LANG_FRENCH;
+export type GsLang = GsLangUser | typeof LANG_SYSTEM;
 
 export function getSystemLanguage() {
     const systemLanguage = navigator.language.split(/[-_]/)[0];
